requirements.txt
click
numpy
pandas
python-dateutil
python-dotenv
pydantic
SQLAlchemy
PyMySQL[rsa]
peewee
mysql-connector-python
matplotlib
PyQt5
openpyxl
pymssql
peewee-mssql